Added header file for math gen class Added a header file for math gen class, modified other sections to work with the new structure.
